fix: Don't use tagless images unless they're pod template names
See https://github.com/jenkins-x/jx/issues/4615. Regardless of what we
do about that issue in the end, we shouldn't be using
`gcr.io/jenkinsxio/builder-go` as the image in the first place. `go`
(i.e., referring to the pod template name) will always pick up the
latest in the version stream, which is what we'd want anyway (and what
we do everywhere else here).

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>